backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
starlette>=0.27.0
asyncio>=3.4.3
pydantic>=2.0.0
pillow>=9.5.0
langgraph>=0.0.26
langchain>=0.1.0
langchain-core>=0.1.14
openai>=1.10.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
python-dotenv>=1.0.0 